pyproject.toml
[project]
name = "concall-transcription"
version = "0.1.0"
description = "Real-Time Indian Concall Transcription & Insight Streaming"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    # Core
    "python-dotenv>=1.0.0",
    # FastAPI & Streaming
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sse-starlette>=2.0.0",
    "python-multipart>=0.0.6",
    # Audio processing
    "pydub>=0.25.1",
    "soundfile>=0.12.1",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    # CLI & Formatting
    "rich>=13.7.0",
    "typer[all]>=0.9.0",
    "faster-whisper>=1.2.1",
    "google-genai>=1.62.0",
]

[project.optional-dependencies]
# ASR Options - uncomment your choice in the container or install manually
whisper = [
    "faster-whisper>=1.0.0",
    "torch>=2.0.0",
]
faster-whisper = [
    "faster-whisper>=1.0.0",
]

# LLM Options - uncomment your choice
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.18.0",
]
google = [
    "google-generativeai>=0.4.0",
]

# Development dependencies
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
]

[project.scripts]
concall = "main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
